Moved class mover RPC handlers
