+++ lib/Container/RpcExtension.php
@@ -10,9 +10,9 @@ use Phpactor\Rpc\Handler\StatusHandler;
 use Phpactor\Extension\WorseReflection\Rpc\GotoDefinitionHandler;
 use Phpactor\Rpc\Handler\CompleteHandler;
 use Phpactor\Rpc\Handler\ClassSearchHandler;
-use Phpactor\Rpc\Handler\ClassCopyHandler;
-use Phpactor\Rpc\Handler\ClassMoveHandler;
-use Phpactor\Rpc\Handler\ReferencesHandler;
+use Phpactor\Extension\ClassMover\Rpc\ClassCopyHandler;
+use Phpactor\Extension\ClassMover\Rpc\ClassMoveHandler;
+use Phpactor\Extension\ClassMover\Rpc\ReferencesHandler;
 use Phpactor\Rpc\Handler\OffsetInfoHandler;
 use Phpactor\Rpc\Handler\TransformHandler;
 use Phpactor\Rpc\Handler\ClassNewHandler;
