+++ lib/Extension/ClassMover/Rpc/ClassCopyHandler.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Phpactor\Rpc\Handler;
+namespace Phpactor\Extension\ClassMover\Rpc;
 
 use Phpactor\Extension\ClassMover\Application\ClassCopy;
 use Phpactor\Rpc\Response\OpenFileResponse;
 use Phpactor\Rpc\Response\Input\TextInput;
 use Phpactor\Application\Logger\NullLogger;
+use Phpactor\Rpc\Handler\AbstractHandler;
 
 class ClassCopyHandler extends AbstractHandler
 {
