+++ tests/Unit/Rpc/Handler/CopyFileHandlerTest.php
@@ -5,7 +5,7 @@ namespace Phpactor\Tests\Unit\Rpc\Handler;
 use Phpactor\Rpc\Handler;
 use Phpactor\Extension\ClassMover\Application\ClassCopy;
 use Phpactor\Rpc\Response\InputCallbackResponse;
-use Phpactor\Rpc\Handler\ClassCopyHandler;
+use Phpactor\Extension\ClassMover\Rpc\ClassCopyHandler;
 use Phpactor\Rpc\Request;
 use Prophecy\Argument;
 use Phpactor\Rpc\Response\OpenFileResponse;
